# api-design/autogenerated/openapi.yaml
openapi: 3.0.0
info:
  title: TodoDoo API
  version: 0.0.0
...
paths:
  /api/auth/login:
    post:
      operationId: Auth_login
      description: Login with email and password
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      security:
        - {}
  /api/auth/logout:
    post:
      operationId: Auth_logout
      description: Logout user by clearing refresh token cookie
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Authentication
      security:
        - {}
  /api/auth/me:
    get:
      operationId: Auth_getCurrentUser
      description: Get current authenticated user
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Authentication
  /api/auth/refresh:
    post:
      operationId: Auth_refreshToken
      description: Refresh access token using refresh token from httpOnly cookie
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Authentication
      security:
        - {}
  /api/auth/register:
    post:
      operationId: Auth_register
      description: Register a new user
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      security:
        - {}
...
security:
  - BearerAuth: []
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      description: Error response
...
    Token:
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          description: Token type (always 'bearer')
      description: JWT token response
...
    User:
      type: object
      required:
        - id
        - email
        - is_active
        - date_created
        - date_updated
      properties:
        id:
          type: string
          description: Unique identifier for the user
        email:
          type: string
          description: User's email address
        is_active:
          type: boolean
          description: Whether the user account is active
        date_created:
          type: string
          format: date-time
          description: When the user was created
        date_updated:
          type: string
          format: date-time
          description: When the user was last updated
      description: User model
    UserLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: User's email address
        password:
          type: string
          description: User's password
      description: Request body for user login
    UserRegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: User's email address
        password:
          type: string
          description: User's password
      description: Request body for user registration
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: http://localhost:8000
    description: Development server
    variables: {}